Restructure GL flavors parsing and module structure
