+++ src/gardenlinux/flavors/__init__.py
@@ -1,0 +1,1 @@
+from .parser import Parser
