+++ tests/test_push_image.py
@@ -1,5 +1,3 @@
-from idlelib.window import registry
-
 import pytest
 import os
